requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-dotenv==1.4.2
django-fontawesome-5==1.0.18
django-heroku==0.3.1
django-manifest-loader==0.2.0
django-rest-auth==0.9.5
django-storages==1.11.1
djangorestframework==3.11.1
djangorestframework-api-key==2.0.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==8.1.0
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.4
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
whitenoise==5.0.1
